# openapi/egdata.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "egdata.app API",
    "version": "0.1.0-alpha.0",
    "description": "REST API for Epic Games Store catalog, pricing, search, and public egdata.app data. This contract starts with public stable REST endpoints and will expand as legacy routes are classified.",
    "contact": {
      "name": "egdata.app",
      "url": "https://egdata.app"
    }
  },
  "servers": [
    {
      "url": "https://api.egdata.app",
      "description": "Production"
    },
    {
      "url": "https://api-gcp.egdata.app",
      "description": "Production GCP mirror"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/offers": {
      "get": {
        "operationId": "listOffers",
        "tags": [
          "Offers"
        ],
        "summary": "List offers with regional price data",
        "parameters": [
          {
            "$ref": "#/components/parameters/country"
          },
          {
            "$ref": "#/components/parameters/locale"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{id}": {
      "get": {
        "operationId": "getOffer",
        "tags": [
          "Offers"
        ],
        "summary": "Get an offer by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          },
          {
            "$ref": "#/components/parameters/locale"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/offers/{id}/assets": {
      "get": {
        "operationId": "listOfferAssets",
        "tags": [
          "Offer Details"
        ],
        "summary": "List assets for an offer",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{id}/items": {
      "get": {
        "operationId": "listOfferItems",
        "tags": [
          "Offer Details"
        ],
        "summary": "List items for an offer",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/offers/{id}/achievements": {
      "get": {
        "operationId": "listOfferAchievements",
        "tags": [
          "Offer Details"
        ],
        "summary": "List achievement sets for an offer",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementSet"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/offers/{id}/builds": {
      "get": {
        "operationId": "listOfferBuilds",
        "tags": [
          "Offer Details"
        ],
        "summary": "List recent builds for an offer",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Build"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EpicSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "egdata.app browser session backed by Epic authentication."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Endpoint-specific API key."
      }
    },
    "parameters": {
      "country": {
        "name": "country",
        "in": "query",
        "required": false,
        "description": "ISO country code used to resolve an Epic Games Store pricing region. Defaults to the EGDATA_COUNTRY cookie, then US.",
        "schema": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2,
          "example": "US"
        }
      },
      "locale": {
        "name": "locale",
        "in": "query",
        "required": false,
        "description": "Exact BCP-47-style locale used to overlay localized public offer text fields. Defaults to en-US.",
        "schema": {
          "type": "string",
          "example": "es-ES"
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records to return. Most list endpoints cap this server-side.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 10
        }
      },
      "page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "One-based page number.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}